+++ pymongo/mongo_client.py
@@ -1512,6 +1512,46 @@ class MongoClient(common.BaseObject):
 
         return self[self.__default_database_name]
 
+    def get_database(self, name, codec_options=None,
+                     read_preference=None, write_concern=None):
+        """Get a :class:`~pymongo.database.Database` with the given name and
+        options.
+
+        Useful for creating a :class:`~pymongo.database.Database` with
+        different codec options, read preference, and/or write concern from
+        this :class:`MongoClient`.
+
+          >>> from pymongo import ReadPreference
+          >>> client.read_preference == ReadPreference.PRIMARY
+          True
+          >>> db1 = client.test
+          >>> db1.read_preference == ReadPreference.PRIMARY
+          True
+          >>> db2 = client.get_database(
+          ...     'test', read_preference=ReadPreference.SECONDARY)
+          >>> db2.read_preference == ReadPreference.SECONDARY
+          True
+
+        :Parameters:
+          - `name`: The name of the database - a string.
+          - `codec_options` (optional): An instance of
+            :class:`~bson.codec_options.CodecOptions`. If ``None`` (the
+            default) the :attr:`codec_options` of this :class:`MongoClient` is
+            used.
+          - `read_preference` (optional): The read preference to use. If
+            ``None`` (the default) the :attr:`read_preference` of this
+            :class:`MongoClient` is used. See :mod:`~pymongo.read_preferences`
+            for options.
+          - `write_concern` (optional): An instance of
+            :class:`~pymongo.write_concern.WriteConcern`. If ``None`` (the
+            default) the :attr:`write_concern` of this :class:`MongoClient` is
+            used.
+
+        .. versionadded:: 2.9
+        """
+        return database.Database(
+            self, name, codec_options, read_preference, write_concern)
+
     @property
     def is_locked(self):
         """Is this server locked? While locked, all write operations