+++ test/test_collection.py
@@ -28,9 +28,10 @@ from nose.plugins.skip import SkipTest
 
 sys.path[0:0] = [""]
 
-from bson.binary import Binary
+from bson.binary import Binary, JAVA_LEGACY
 from bson.regex import Regex
 from bson.code import Code
+from bson.codec_options import CodecOptions
 from bson.dbref import DBRef
 from bson.objectid import ObjectId
 from bson.py3compat import b
@@ -40,9 +41,6 @@ from pymongo import (ASCENDING, DESCENDING, GEO2D,
 from pymongo import message as message_module
 from pymongo.collection import Collection
 from pymongo.command_cursor import CommandCursor
-from pymongo.mongo_replica_set_client import MongoReplicaSetClient
-from pymongo.read_preferences import ReadPreference
-from pymongo.son_manipulator import SONManipulator
 from pymongo.errors import (DocumentTooLarge,
                             DuplicateKeyError,
                             InvalidDocument,
@@ -50,6 +48,10 @@ from pymongo.errors import (DocumentTooLarge,
                             InvalidOperation,
                             OperationFailure,
                             WTimeoutError)
+from pymongo.mongo_replica_set_client import MongoReplicaSetClient
+from pymongo.read_preferences import ReadPreference, Secondary
+from pymongo.son_manipulator import SONManipulator
+from pymongo.write_concern import WriteConcern
 from test.test_client import get_client
 from test.utils import (catch_warnings, enable_text_search,
                         get_pool, is_mongos, joinall, oid_generated_on_client)
@@ -108,6 +110,26 @@ class TestCollection(unittest.TestCase):
 
         # No exception
         self.db.drop_collection('test')
+
+    def test_with_options(self):
+        coll = self.db.test
+        codec_options = CodecOptions(
+            tz_aware=True, uuid_representation=JAVA_LEGACY)
+        write_concern = WriteConcern(w=2, j=True)
+        coll2 = coll.with_options(
+            codec_options, ReadPreference.SECONDARY, write_concern)
+        self.assertEqual(codec_options, coll2.codec_options)
+        self.assertEqual(JAVA_LEGACY, coll2.uuid_subtype)
+        self.assertEqual(ReadPreference.SECONDARY, coll2.read_preference)
+        self.assertEqual(write_concern.document, coll2.write_concern)
+
+        pref = Secondary([{"dc": "sf"}])
+        coll2 = coll.with_options(read_preference=pref)
+        self.assertEqual(pref.mode, coll2.read_preference)
+        self.assertEqual(pref.tag_sets, coll2.tag_sets)
+        self.assertEqual(coll.codec_options, coll2.codec_options)
+        self.assertEqual(coll.uuid_subtype, coll2.uuid_subtype)
+        self.assertEqual(coll.write_concern, coll2.write_concern)
 
     def test_create_index(self):
         db = self.db