+++ test/test_cursor.py
@@ -686,8 +686,8 @@ class TestCursor(unittest.TestCase):
         self.assertEqual(cursor._Cursor__skip, cursor2._Cursor__skip)
         self.assertEqual(cursor._Cursor__limit, cursor2._Cursor__limit)
         self.assertEqual(cursor._Cursor__snapshot, cursor2._Cursor__snapshot)
-        self.assertEqual(type(cursor._Cursor__as_class),
-                         type(cursor2._Cursor__as_class))
+        self.assertEqual(type(cursor._Cursor__codec_options),
+                         type(cursor2._Cursor__codec_options))
         self.assertEqual(cursor._Cursor__slave_okay,
                          cursor2._Cursor__slave_okay)
         self.assertEqual(cursor._Cursor__manipulate,