+++ test/test_master_slave_connection.py
@@ -471,6 +471,7 @@ class TestMasterSlaveConnection(unittest.TestCase, TestRequestMixin):
         self.assertFalse(isinstance(db.test.find_one(), SON))
 
         c.document_class = SON
+        db = c.pymongo_test
 
         self.assertEqual(SON, c.document_class)
         self.assertTrue(isinstance(db.test.find_one(), SON))
@@ -484,6 +485,7 @@ class TestMasterSlaveConnection(unittest.TestCase, TestRequestMixin):
         self.assertFalse(isinstance(db.test.find_one(as_class=dict), SON))
 
         c.document_class = dict
+        db = c.pymongo_test
 
         self.assertEqual(dict, c.document_class)
         self.assertTrue(isinstance(db.test.find_one(), dict))